refactor: extract task tools into lithos.tools; delete _register_tools
Final extraction slice — all 17 lithos_task_* tools move to
tools/tasks.py along with _serialize_task_record. The five verbatim
'except CoordinationError -> envelope' blocks are replaced by
tool_span(map_coordination_error=True) on every task tool (the seam
built for exactly this in PR 3); the mapping is uniform across the
domain now instead of hand-placed per call.

_register_tools is deleted; register_all(self.mcp, self) is the single
registration path. server.py drops to ~1,015 lines: lifecycle wiring,
HTTP/SSE/health/audit endpoints, the event-emit helper, and the
task-feedback helpers (whose relocation to CognitiveMemory is a
separate review task).

All 37 tools now live in lithos/tools/; the envelope-conformance suite,
frozen tool snapshot, and AST pairing check pass unmodified.

Part of the 2026-07 architecture review (tracker task 9f2be54f, PR 5 of 6).
